+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/cds-canvas.component.ts
@@ -16,7 +16,7 @@ import { Intent, Form } from 'src/app/models/intent-model';
 import { Button, Action} from 'src/app/models/action-model';
 
 // UTILS //
-import { INTENT_COLORS, RESERVED_INTENT_NAMES, TYPE_INTENT_ELEMENT, TYPE_OF_MENU, INTENT_TEMP_ID, OPTIONS } from '../../utils';
+import { INTENT_COLORS, RESERVED_INTENT_NAMES, TYPE_INTENT_ELEMENT, TYPE_OF_MENU, INTENT_TEMP_ID, OPTIONS, STAGE_SETTINGS } from '../../utils';
 import { LOGOS_ITEMS } from './../../utils-resources';
 
 
@@ -26,6 +26,7 @@ import { LoggerService } from 'src/chat21-core/providers/abstract/logger.service
 import { TYPE_ACTION } from 'src/app/chatbot-design-studio/utils-actions';
 import { AppStorageService } from 'src/chat21-core/providers/abstract/app-storage.service';
 import { storage } from 'firebase';
+import { LogService } from 'src/app/services/log.service';
 
 // const swal = require('sweetalert');
 
@@ -43,7 +44,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   @Input() onHeaderTestItOut: Observable<Intent>
 
   /** listners */
-  listenerConnectorDrawn: (e: CustomEvent) => void;
+  listnerConnectorDrawn: (e: CustomEvent) => void;
   listnerMovedAndScaled: (e: CustomEvent) => void;
   listnerKeydown: (e: any) => void;
   listnerConnectorSelected: (e: CustomEvent) => void;
@@ -105,6 +106,10 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   testitOutFirstClick: boolean = false;
   IS_OPEN_PANEL_WIDGET: boolean = false;
 
+  /** panel widget loaded */
+  private subscriptionWidgetLoaded: Subscription;
+
+
   /** panel options */
   private subscriptionUndoRedo: Subscription;
   stateUndoRedo: any = {undo:false, redo: false};
@@ -122,6 +127,10 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   IS_OPEN_COLOR_MENU: boolean = false;
   positionColortMenu: any = { 'x': 0, 'y': 0 };
 
+  /** panel widget log */
+  IS_OPEN_WIDGET_LOG: boolean = false;
+  
+
   private logger: LoggerService = LoggerInstance.getInstance()
   
   IS_OPEN_PANEL_INTENT_DETAIL: boolean = false;
@@ -129,15 +138,16 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
 
 
   constructor(
-    private intentService: IntentService,
-    private stageService: StageService,
-    private connectorService: ConnectorService,
-    private controllerService: ControllerService,
-    private translate: TranslateService,
-    public dashboardService: DashboardService,
-    private changeDetectorRef: ChangeDetectorRef,
-    private route: ActivatedRoute, 
-    public appStorageService: AppStorageService
+    private readonly intentService: IntentService,
+    private readonly stageService: StageService,
+    private readonly connectorService: ConnectorService,
+    private readonly controllerService: ControllerService,
+    // private translate: TranslateService,
+    public readonly dashboardService: DashboardService,
+    private readonly changeDetectorRef: ChangeDetectorRef,
+    private readonly route: ActivatedRoute, 
+    public appStorageService: AppStorageService,
+    public logService: LogService
   ) {
     this.setSubscriptions();
     this.setListnerEvents();
@@ -167,13 +177,16 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     if (this.subscriptionOpenWidgetPanel) {
       this.subscriptionOpenWidgetPanel.unsubscribe();
     }
-    if (this.subscriptionUndoRedo) {
-      this.subscriptionUndoRedo.unsubscribe();
+    if (this.subscriptionWidgetLoaded) {
+      this.subscriptionWidgetLoaded.unsubscribe();
     }
     if (this.subscriptionUndoRedo) {
       this.subscriptionUndoRedo.unsubscribe();
     }
-    document.removeEventListener("connector-drawn", this.listenerConnectorDrawn, false);
+    if (this.subscriptionUndoRedo) {
+      this.subscriptionUndoRedo.unsubscribe();
+    }
+    document.removeEventListener("connector-drawn", this.listnerConnectorDrawn, false);
     document.removeEventListener("moved-and-scaled", this.listnerMovedAndScaled, false);
     document.removeEventListener("start-dragging", this.listnerStartDragging, false);
     document.removeEventListener("keydown", this.listnerKeydown, false);
@@ -191,11 +204,15 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   ngAfterViewInit() {
     this.logger.log("[CDS-CANVAS]  •••• ngAfterViewInit ••••");
     this.stageService.initializeStage(this.id_faq_kb);
-    // this.stageService.initStageSettings(this.id_faq_kb);
+    if(this.stageService.settings?.open_intent_list_state != null){
+      this.IS_OPEN_INTENTS_LIST = this.stageService.settings.open_intent_list_state;
+    }
     
+    // this.stageService.initStageSettings(this.id_faq_kb);
     this.stageService.setDrawer();
     this.connectorService.initializeConnectors();
     this.changeDetectorRef.detectChanges();
+    
     setTimeout(() => {
       this.showStageForLimitTime();
     }, 2000);
@@ -237,13 +254,13 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   onIntentRendered(intentID) {
     if(this.stageService.loaded === false && this.renderedAllElements === false){
       this.labelInfoLoading = 'CDSCanvas.intentsProgress';
-      if(this.mapOfIntents[intentID] && this.mapOfIntents[intentID].shown === false) {
-        this.mapOfIntents[intentID].shown = true;
+      if(this.mapOfIntents[intentID]) { //&& this.mapOfIntents[intentID].shown === false
+        this.mapOfIntents[intentID].shown = 'true';
         this.countRenderedElements++;
         this.loadingProgress += (this.countRenderedElements/this.totElementsOnTheStage)*100;
       }
       this.logger.log("[CDS-CANVAS3] •••• onIntentRendered •••• ", intentID, this.countRenderedElements);
-      const allShownTrue = Object.values(this.mapOfIntents).every(intent => intent.shown === true);
+      const allShownTrue = Object.values(this.mapOfIntents).every(intent => intent.shown === 'true');
       if(allShownTrue){ 
         this.onAllIntentsRendered();
       }
@@ -258,11 +275,11 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   }
 
   checkAllConnectors(connector){
-    // this.logger.log("[CDS-CANVAS3]  •••• checkAllConnectors ••••", connector);
+    // this.logger.log("[CDS-CANVAS]  •••• checkAllConnectors ••••", connector.id);
     if(this.stageService.loaded === false && this.renderedAllElements === false){
       this.labelInfoLoading = 'CDSCanvas.connectorsProgress';
-      if(this.mapOfConnectors[connector.id] && this.mapOfConnectors[connector.id].shown === false) {
-        this.mapOfConnectors[connector.id].shown = true;
+      if(this.mapOfConnectors[connector.id] && this.mapOfConnectors[connector.id].shown !== 'true') {
+        this.mapOfConnectors[connector.id].shown = 'true';
         this.countRenderedElements++;
         this.loadingProgress += (this.countRenderedElements/this.totElementsOnTheStage)*100;
         this.logger.log("[CDS-CANVAS]  •••• E' stato creato un nuovo connettore verifico ••••", connector.id, this.countRenderedElements);
@@ -273,8 +290,8 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
 
   private checkAndShowStage(){
     if(this.stageService.loaded === false){
-      const allShownTrue = Object.values(this.mapOfConnectors).every(connector => connector.shown == true);
-      this.logger.log("[CDS-CANVAS]  •••• checkAndshowStage", this.mapOfConnectors, allShownTrue);
+      const allShownTrue = Object.values(this.mapOfConnectors).every(connector => connector.shown !== 'false');
+      // this.logger.log("[CDS-CANVAS]  •••• checkAndshowStage", this.mapOfConnectors, allShownTrue);
       if(allShownTrue){ 
         this.stageService.loaded = true;
         this.loadingProgress = 100;
@@ -421,13 +438,20 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       }
     });    
 
+    /** SUBSCRIBE TO THE LOADED WIDGET */
+    this.subscriptionWidgetLoaded = this.logService.BSWidgetLoaded.subscribe((event: any) => {
+      this.logger.log("[CDS-CANVAS] logService loaded ", event);
+      if(event){
+        this.IS_OPEN_WIDGET_LOG = true;
+      }
+    });    
   }
 
    /** initialize */
    private async initialize(){
     this.id_faq_kb = this.dashboardService.id_faq_kb;
     this.listOfIntents = [];
-    const getAllIntents = await this.intentService.getAllIntents(this.id_faq_kb);
+    let getAllIntents = await this.intentService.getAllIntents(this.id_faq_kb);
     if (getAllIntents) {
       this.listOfIntents = this.intentService.listOfIntents;
       // // this.initListOfIntents();
@@ -439,7 +463,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       const numIntents = Object.values(this.mapOfIntents).length;
       const numConnectors = Object.values(this.mapOfConnectors).length;
       this.totElementsOnTheStage = numIntents+numConnectors;
-      this.logger.log('[CDS-CANVAS3] totElementsOnTheStage ::', this.stageService.loaded, numIntents, numConnectors);
+      this.logger.log('[CDS-CANVAS] totElementsOnTheStage ::', this.stageService.loaded, numIntents, numConnectors);
       // scaleAndcenterStageOnCenterPosition(this.listOfIntents)
     }
     this.subscriptionOpenWidgetPanel = this.intentService.BStestiTout.pipe(skip(1)).subscribe((event) => this.onTestItOut(event));
@@ -458,6 +482,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   /** closeAllPanels */
   private closeAllPanels(){
     this.IS_OPEN_PANEL_WIDGET = false;
+    this.IS_OPEN_WIDGET_LOG = false;
     this.IS_OPEN_PANEL_ACTION_DETAIL = false;
     this.IS_OPEN_PANEL_INTENT_DETAIL = false;
     this.IS_OPEN_PANEL_BUTTON_CONFIG = false;
@@ -507,12 +532,12 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   setListnerEvents(){
     /** LISTENER OF TILEDESK STAGE */
     /** triggers when a connector is drawn on the stage  */
-    this.listenerConnectorDrawn = (e: CustomEvent) => {
-    const connector = e.detail.connector;
-    this.setConnectorColor(connector);
-    this.checkAllConnectors(connector);
-  };
-  document.addEventListener("connector-drawn", this.listenerConnectorDrawn, false);
+    this.listnerConnectorDrawn = (e: CustomEvent) => {
+      const connector = e.detail.connector;
+      this.setConnectorColor(connector);
+      this.checkAllConnectors(connector);
+    };
+    document.addEventListener("connector-drawn", this.listnerConnectorDrawn, false);
     
 
     
@@ -592,11 +617,18 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     */
     this.listnerConnectorDraftReleased = (e: CustomEvent) => {
       this.logger.log("[CDS-CANVAS] connector-draft-released :: ", e.detail);
-      if(!e || !e.detail) return;
-      const detail = e.detail;
+      if(!e?.detail) return;
+      let detail = e.detail;
       const arrayOfClass = detail.target.classList.value.split(' ');
       if (detail.target && arrayOfClass.includes("receiver-elements-dropped-on-stage") && detail.toPoint && detail.menuPoint) {
         this.logger.log("[CDS-CANVAS] ho rilasciato il connettore tratteggiato nello stage (nell'elemento con classe 'receiver_elements_dropped_on_stage') e quindi apro il float menu");
+        const intentId = e.detail.fromId.split('/')[0];
+        const intent = this.intentService.getIntentFromId(intentId);
+        if(intent.attributes?.color){
+          detail.color = intent.attributes.color;
+        } else {
+          detail.color = INTENT_COLORS.COLOR1;
+        }
         this.openFloatMenuOnConnectorDraftReleased(detail);
       } else {
         this.logger.log("[CDS-CANVAS] ho rilasciato in un punto qualsiasi del DS ma non sullo stage quindi non devo aprire il menu", detail);
@@ -702,7 +734,6 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       }
     };
     document.addEventListener("keydown", this.listnerKeydown, false);
-
   }
 
   // ---------------------------------------------------------
@@ -887,6 +918,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     this.logger.log('[CDS-CANVAS] onToogleSidebarIntentsList  ')
     this.IS_OPEN_INTENTS_LIST = !this.IS_OPEN_INTENTS_LIST;
     this.removeConnectorDraftAndCloseFloatMenu();
+    this.stageService.saveSettings(this.id_faq_kb, STAGE_SETTINGS.openIntentListState, this.IS_OPEN_INTENTS_LIST);
     this.logger.log('[CDS-CANVAS] onToogleSidebarIntentsList   this.IS_OPEN_INTENTS_LIST ',  this.IS_OPEN_INTENTS_LIST)
   }
 
@@ -937,7 +969,8 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   async createNewIntentFromConnectorDraft(typeAction, connectorDraft){
     const toPoint = connectorDraft.toPoint;
     const newAction = this.intentService.createNewAction(typeAction);
-    let intent = this.intentService.createNewIntent(this.id_faq_kb, newAction, toPoint);
+    let intent = this.intentService.createNewIntent(this.id_faq_kb, newAction, toPoint, connectorDraft.color);
+    this.logger.log('[CDS-CANVAS] ho creato intent: ', intent);
     this.intentService.addNewIntentToListOfIntents(intent);
     this.removeConnectorDraftAndCloseFloatMenu();
     const fromId = connectorDraft.fromId;
@@ -953,8 +986,8 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       let pos = this.listOfIntents.length-1;
       this.logger.log('[CDS-CANVAS] sto per chiamare settingAndSaveNewIntent ', prevIntent, nowIntent);
       this.settingAndSaveNewIntent(pos, intent, nowIntent, prevIntent);
-      // this.logger.log("[CDS-CANVAS] sto per aggiornare l'intent ", nowIntent);
-      // this.updateIntent(nowIntent, 0, false);
+      // //this.logger.log("[CDS-CANVAS] sto per aggiornare l'intent ", nowIntent);
+      // //this.updateIntent(nowIntent, 0, false);
     }
   }
 
@@ -1230,30 +1263,26 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   // - button configuration panel  
   // -------------------------------------------------------
   onTestItOut(intent: Intent) {
-    // this.testItOut.emit(true);
+    // // this.testItOut.emit(true);
+    this.closeAllPanels();
     this.testitOutFirstClick = true;
     this.logger.log('[CDS-CANVAS] onTestItOut intent ', intent);
-    this.IS_OPEN_PANEL_WIDGET = true
-    this.intentService.startTestWithIntent(intent)
-    // if(typeof event === "boolean"){
-    //   this.IS_OPEN_PANEL_WIDGET = true;
-    // } else {
-    //   this.IS_OPEN_PANEL_WIDGET = !this.IS_OPEN_PANEL_WIDGET;
-    // }
-    if(this.IS_OPEN_PANEL_WIDGET){
-      this.controllerService.closeActionDetailPanel();
-      this.controllerService.closeButtonPanel();
-      // this.intentService.setLiveActiveIntent(null);
-      this.controllerService.closeAddActionMenu();
-      this.connectorService.removeConnectorDraft();
-    }
+    this.intentService.startTestWithIntent(intent);
+    this.controllerService.closeActionDetailPanel();
+    this.controllerService.closeButtonPanel();
+    // // this.intentService.setLiveActiveIntent(null);
+    this.controllerService.closeAddActionMenu();
+    this.connectorService.removeConnectorDraft();
     if(intent){
-      // this.intentSelected = intent;
+      // // this.intentSelected = intent;
       this.intentService.setIntentSelectedById(intent.intent_id);
       this.intentService.setIntentSelected(intent.intent_id);
       this.closeExtraPanels();
     }
 
+    setTimeout(() => {
+      this.IS_OPEN_PANEL_WIDGET = true;
+    }, 500);
   }
 
   /** onActionDeleted */
@@ -1355,18 +1384,18 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     this.IS_OPEN_ADD_ACTIONS_MENU = true;
     const connectorDraft = this.connectorService.connectorDraft;
 
-    if (connectorDraft && connectorDraft.toPoint && !this.hasClickedAddAction) {
-      this.logger.log("[CDS-CANVAS] ho trascinato il connettore e sto per creare un intent");
+    if (connectorDraft?.toPoint && !this.hasClickedAddAction) {
+      this.logger.log("[CDS-CANVAS] ho trascinato il connettore e sto per creare un intent", connectorDraft);
       this.createNewIntentFromConnectorDraft(event.type, connectorDraft);
-      // this.removeConnectorDraftAndCloseFloatMenu();
+      // // this.removeConnectorDraftAndCloseFloatMenu();
     } 
     else if (this.hasClickedAddAction) {
       this.logger.log("[CDS-CANVAS] ho premuto + quindi creo una nuova action e la aggiungo all'intent");
       const newAction = this.intentService.createNewAction(event.type);
-      // this.intentSelected.actions.push(newAction);
+      // // this.intentSelected.actions.push(newAction);
       this.intentService.addActionToIntentSelected(newAction);
-      // this.intentService.updateIntentSelected();
-      // this.updateIntent(this.intentSelected, 0, true);
+      // // this.intentService.updateIntentSelected();
+      // // this.updateIntent(this.intentSelected, 0, true);
       this.controllerService.closeAddActionMenu();
     }