+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/cds-canvas.module.ts
@@ -35,6 +35,8 @@ import { PreloadBarComponent } from 'src/app/chatbot-design-studio/cds-base-elem
 import { ImagePreloaderResolver } from 'src/app/resolver/image.resolver';
 import { CdsColorMenuComponent } from './base-elements/cds-color-menu/cds-color-menu.component';
 import { CdsConnectorInComponent } from './base-elements/cds-connector-in/cds-connector-in.component';
+import { CdsWidgetLogsComponent } from './base-elements/cds-widget-logs/cds-widget-logs.component';
+import { ChangeAlphaColorComponent } from 'src/app/modals/change-alpha-color/change-alpha-color.component';
 import { CdsActionWebhookComponent } from './actions/list/cds-action-webhook/cds-action-webhook.component';
 
 const routes: Routes = [
@@ -91,6 +93,7 @@ const routes: Routes = [
 
     //MODALS
     MouseTipsComponent,
+    ChangeAlphaColorComponent,
 
     //EVENT BASE COMPONENT
     CdsEventComponent,
@@ -99,7 +102,6 @@ const routes: Routes = [
     PreloadBarComponent,
     CdsColorMenuComponent,
     CdsConnectorInComponent,
-
     CdsActionWebhookComponent
   ],
   imports: [