+++ src/app/chatbot-design-studio/services/intent.service.ts
@@ -3,7 +3,7 @@ import { Subject, BehaviorSubject } from 'rxjs';
 import { LoggerService } from 'src/chat21-core/providers/abstract/logger.service';
 import { v4 as uuidv4 } from 'uuid';
 
-import { ActionReply, ActionAgent, ActionAssignFunction, ActionAssignVariable, ActionChangeDepartment, ActionClose, ActionDeleteVariable, ActionEmail, ActionHideMessage, ActionIntentConnected, ActionJsonCondition, ActionOnlineAgent, ActionOpenHours, ActionRandomReply, ActionReplaceBot, ActionWait, ActionWebRequest, Command, Wait, Message, Expression, Action, ActionAskGPT, ActionWhatsappAttribute, ActionWhatsappStatic, ActionWebRequestV2, ActionGPTTask, ActionCaptureUserReply, ActionQapla, ActionCondition, ActionMake, ActionAssignVariableV2, ActionHubspot, ActionCode, ActionReplaceBotV2, ActionAskGPTV2, ActionCustomerio, ActionVoice, ActionBrevo, Attributes, ActionN8n, ActionGPTAssistant, ActionReplyV2, ActionOnlineAgentV2, ActionLeadUpdate, ActionClearTranscript, ActionMoveToUnassigned, ActionConnectBlock, ActionAddTags, ActionSendWhatsapp, WhatsappBroadcast, ActionReplaceBotV3, ActionAiPrompt, ActionWebhook } from 'src/app/models/action-model';
+import { ActionReply, ActionAgent, ActionAssignFunction, ActionAssignVariable, ActionChangeDepartment, ActionClose, ActionDeleteVariable, ActionEmail, ActionHideMessage, ActionIntentConnected, ActionJsonCondition, ActionOnlineAgent, ActionOpenHours, ActionRandomReply, ActionReplaceBot, ActionWait, ActionWebRequest, Command, Wait, Message, Expression, Action, ActionAskGPT, ActionWhatsappAttribute, ActionWhatsappStatic, ActionWebRequestV2, ActionGPTTask, ActionCaptureUserReply, ActionQapla, ActionCondition, ActionMake, ActionAssignVariableV2, ActionHubspot, ActionCode, ActionReplaceBotV2, ActionAskGPTV2, ActionCustomerio, ActionVoice, ActionBrevo, Attributes, ActionN8n, ActionGPTAssistant, ActionReplyV2, ActionOnlineAgentV2, ActionLeadUpdate, ActionClearTranscript, ActionMoveToUnassigned, ActionConnectBlock, ActionAddTags, ActionSendWhatsapp, WhatsappBroadcast, ActionReplaceBotV3, ActionAiPrompt, ActionWebhook, ActionWebRespose } from 'src/app/models/action-model';
 import { Intent } from 'src/app/models/intent-model';
 import { FaqService } from 'src/app/services/faq.service';
 import { FaqKbService } from 'src/app/services/faq-kb.service';
@@ -62,7 +62,6 @@ export class IntentService {
   operationsUndo: any = [];
   operationsRedo: any = [];
   // newPosition: any = {'x':0, 'y':0};
-  
 
   private changedConnector = new Subject<any>();
   public isChangedConnector$ = this.changedConnector.asObservable();
@@ -129,7 +128,7 @@ export class IntentService {
     this.intentSelectedID = null;
     this.intentActive = false;
     if(this.listOfIntents && this.listOfIntents.length > 0){
-      let startIntent = this.listOfIntents.filter(obj => ( obj.intent_display_name.trim() === TYPE_INTENT_NAME.DISPLAY_NAME_START));
+      let startIntent = this.listOfIntents.filter(obj => ( obj.intent_display_name.trim() === TYPE_INTENT_NAME.START));
       // this.logger.log('setDefaultIntentSelected: ', startIntent, startIntent[0]);
       if(startIntent && startIntent.length>0){
         this.intentSelected = startIntent[0];
@@ -179,7 +178,7 @@ export class IntentService {
   }
 
   public setIntentSelectedPosition(x, y){
-    if (this.intentSelected && this.intentSelected.attributes) {
+    if (this.intentSelected?.attributes) {
       if (!this.intentSelected.attributes.position) {
         this.intentSelected.attributes.position = {};
       }
@@ -342,14 +341,15 @@ export class IntentService {
  
 
   /** create a new intent when drag an action on the stage */
-  public createNewIntent(id_faq_kb: string, action: any, pos:any){
+  public createNewIntent(id_faq_kb: string, action: any, pos:any, color?: string){
     let intent = new Intent();
     const chatbot_id = this.dashboardService.id_faq_kb;
     intent.id_faq_kb = chatbot_id;
     intent.attributes.position = pos;
     intent.intent_display_name = this.setDisplayName();
-    // //let actionIntent = this.createNewAction(TYPE_ACTION.INTENT);
-    // //intent.actions.push(actionIntent);
+    if(color){
+      intent.attributes.color = color;
+    }
     intent.actions.push(action);
 
     if(action._tdActionType === TYPE_EVENT_CATEGORY.WEBHOOK){
@@ -769,6 +769,8 @@ export class IntentService {
         return { name: a.intent_display_name, value: '#' + a.intent_id, icon: 'rocket_launch' }
       } else if (a.intent_display_name.trim() === RESERVED_INTENT_NAMES.DEFAULT_FALLBACK) {
         return { name: a.intent_display_name, value: '#' + a.intent_id, icon: 'undo' }
+      } else if (a.intent_display_name.trim() === RESERVED_INTENT_NAMES.CLOSE) {
+        return { name: a.intent_display_name, value: '#' + a.intent_id, icon: 'call_end' }
       } else {
         return { name: a.intent_display_name, value: '#' + a.intent_id, icon: 'label_important_outline' }
       }
@@ -846,7 +848,6 @@ export class IntentService {
         }
         let id_faq_kb = this.dashboardService.id_faq_kb;
         this.stageService.centerStageOnHorizontalPosition(id_faq_kb, startElement, left);
-
       }
     }
   }
@@ -943,6 +944,9 @@ export class IntentService {
       action.assignResultTo= 'result'
       action.assignStatusTo = 'status';
       action.assignErrorTo = 'error';
+    }
+    if(typeAction === TYPE_ACTION.WEB_RESPONSE){
+      action = new ActionWebRespose();
     }
     if(typeAction === TYPE_ACTION.AGENT){
       action = new ActionAgent();
@@ -1766,12 +1770,13 @@ export class IntentService {
     this.logger.log('[INTENT SERVICE] -> pasteElementToStage, ', element, point);
     if(element && element.type === 'INTENT'){
       let newIntent_id = uuidv4();
-      let prevIntent = element.element;
+      let prevIntent = this.replaceId(element.element, '_tdActionId');
       let newAction = prevIntent.actions[0];
       let newIntent = this.createNewIntent(element.chatbot, newAction, 0);
       newIntent.attributes = prevIntent.attributes;
       newIntent.attributes.position = point;
       newIntent.actions = prevIntent.actions;
+      this.logger.log('[INTENT SERVICE] -> prevIntent ', prevIntent);
       this.pasteIntentOntoStage(newIntent, prevIntent.intent_id, newIntent_id);
     } else if(element && element.type === 'ACTION'){
       // let newAction = element.element;
@@ -1824,5 +1829,18 @@ export class IntentService {
     //return results;
   }
 
+
+  private replaceId(obj: any, keyToReplace: string) {
+    for (const key in obj) {
+      if (obj.hasOwnProperty(key)) {
+        if (key === keyToReplace) {
+          obj[key] = uuidv4();
+        } else if (typeof obj[key] === "object" && obj[key] !== null) {
+          this.replaceId(obj[key], keyToReplace);
+        }
+      }
+    }
+    return obj;
+  }
 
 }